data-pipeline/airflow-dags/requirements.txt
# Airflow
apache-airflow[amazon,slack,apache-spark,trino]==2.8.0

# Connectors and Providers
apache-airflow-providers-apache-spark==4.5.0
apache-airflow-providers-trino==5.2.0
apache-airflow-providers-slack==8.3.0
apache-airflow-providers-http==4.5.0
apache-airflow-providers-postgres==5.8.0

# Database and Cache
psycopg2-binary==2.9.9
redis==5.0.1

# AWS
boto3==1.28.85
botocore==1.31.85

# API and Data
trino==0.21.0
requests==2.31.0

# Stripe
stripe==7.4.0

# Utilities
python-dateutil==2.8.2
pytz==2023.3
Babel==2.14.0

# Logging
python-json-logger==2.0.7
